fixed minor bug in tests.
Former-commit-id: 80b5a9664df6af2f5bddce71a040bd622592a411

+++ pyworkflow/em/packages/xmipp3/xmipp3.py
@@ -151,7 +151,7 @@ def findRow(md, label, value):
 
 def findRowById(md, value):
     """ Same as findRow, but using MDL_ITEM_ID for label. """
-    return findRow(md, xmipp.MDL_ITEM_ID, value)
+    return findRow(md, xmipp.MDL_ITEM_ID, long(value))
   
   
 class XmippSet():
